Move request and output_formatter out of rolling_batch
