+++ engines/python/setup/djl_python/rolling_batch/lmi_dist_rolling_batch.py
@@ -19,7 +19,8 @@ from lmi_dist.init_engine import engine_from_args
 from vllm.lora.request import LoRARequest
 from vllm import SamplingParams
 
-from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception, Token, filter_unused_generation_params
+from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception, filter_unused_generation_params
+from djl_python.request_io import Token
 from djl_python.rolling_batch.rolling_batch_vllm_utils import (
     get_speculative_decoding_metrics_record, update_request_cache_with_output,
     supports_speculative_decoding, get_lora_request_params, DTYPE_MAPPER,
