+++ engines/python/setup/djl_python/rolling_batch/vllm_rolling_batch.py
@@ -16,7 +16,8 @@ from collections import OrderedDict, defaultdict
 from vllm import EngineArgs, LLMEngine, SamplingParams
 from vllm.utils import random_uuid
 from vllm.lora.request import LoRARequest
-from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception, Token, filter_unused_generation_params
+from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception, filter_unused_generation_params
+from djl_python.request_io import Token
 from djl_python.rolling_batch.rolling_batch_vllm_utils import (
     update_request_cache_with_output, get_lora_request_params, DTYPE_MAPPER,
     FINISH_REASON_MAPPER)
