+++ engines/python/setup/djl_python/tests/rolling_batch/fake_rolling_batch.py
@@ -15,7 +15,8 @@ from collections import OrderedDict
 from transformers import AutoTokenizer
 
 from djl_python.properties_manager.properties import Properties
-from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception, Token
+from djl_python.rolling_batch.rolling_batch import RollingBatch, stop_on_any_exception
+from djl_python.request_io import Token
 
 
 class FakeRollingBatch(RollingBatch):
